requirements.txt
# Portfolio Dashboard Requirements
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data processing
pandas==2.1.3
numpy==1.25.2
scipy==1.11.4

# Database and API clients
supabase==2.0.2
httpx==0.25.2
aiohttp==3.9.1

# Environment management
python-dotenv==1.0.0

# Data visualization (if needed)
plotly==5.17.0

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
pandas-market-calendars==4.3.3
